3333: Check /ida symlink exists before creating
This post-install script failed if the symlink already existed.

+++ packaging/postinst.sh
@@ -1,4 +1,4 @@
 front scale front=1
 mkdir -p /ida
-ln -s /opt/front /ida/front
+[ ! -L /ida/front ] && ln -s /opt/front /ida/front
 front run rake tmp:create
